Split themes into separate file
